requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
frozenlist==1.3.1
future==0.18.2
idna==3.3
line-bot-sdk==2.3.0
multidict==6.0.2
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
yarl==1.8.1
